+++ lib/cloudscrape_client/api.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require "digest"
 require "faraday"
 require "faraday_middleware"
@@ -21,27 +23,29 @@ class CloudscrapeClient
       new.delete(*args)
     end
 
-    def get(domain:, url:, options: {})
-      connection(domain).get(URI.escape(url), options)
+    def get(domain:, url:, content_type:, options: {})
+      connection(domain: domain, content_type: content_type)
+        .get(URI.escape(url), options)
     end
 
-    def post(domain:, url:, options: {})
-      connection(domain).post do |req|
+    def post(domain:, url:, content_type:, options: {})
+      connection(domain: domain, content_type: content_type).post do |req|
         req.url URI.escape(url)
-        req.headers["Content-Type"] = "application/json"
+        req.headers["Content-Type"] = content_type.to_s
         req.body = options.to_json
       end
     end
 
-    def delete(domain:, url:, options: {})
-      connection(domain).delete(URI.escape(url), options)
+    def delete(domain:, url:, content_type:, options: {})
+      connection(domain: domain, content_type: content_type)
+        .delete(URI.escape(url), options)
     end
 
     private
 
     def access_key
-      fail InvalidAccountId, account_id unless account_id
-      fail InvalidApiKey, api_key unless api_key
+      raise InvalidAccountId, account_id unless account_id
+      raise InvalidApiKey, api_key unless api_key
 
       Digest::MD5.hexdigest(account_id + api_key)
     end
@@ -60,7 +64,7 @@ class CloudscrapeClient
 
     # rubocop:disable Metrics/AbcSize
     # rubocop:disable Metrics/MethodLength
-    def connection(domain)
+    def connection(domain:, content_type:)
       Faraday.new(url: domain) do |faraday|
         faraday.request :url_encoded
 
@@ -72,10 +76,11 @@ class CloudscrapeClient
                         accept: "application/json",
                         "X-CloudScrape-Access" => access_key,
                         "X-CloudScrape-Account" => account_id,
-                        content_type: "application/json"
+                        content_type: content_type.to_s
 
         if CloudscrapeClient.configuration.verbose
-          faraday.response :logger, CloudscrapeClient.configuration.logger
+          faraday.use :extended_logging,
+                      logger: CloudscrapeClient.configuration.logger
         end
 
         faraday.response :multi_json,