+++ lib/cloudscrape_client/dto.rb
@@ -1,8 +1,12 @@
+# frozen_string_literal: true
+
 require "cloudscrape_client/api"
 require "cloudscrape_client/validate"
 
 class CloudscrapeClient
   class DTO
+    DEFAULT_CONTENT_TYPE = MIME::Types["application/json"].first
+
     attr_reader :options
 
     def initialize(options:)
@@ -30,6 +34,7 @@ class CloudscrapeClient
         method,
         domain: domain,
         url: endpoint,
+        content_type: content_type,
         options: {
           api_key: api_key,
           format: "json"
@@ -41,8 +46,12 @@ class CloudscrapeClient
       {}
     end
 
+    def content_type
+      DEFAULT_CONTENT_TYPE
+    end
+
     def endpoint
-      fail NotImplementedError, "Inheriting class must implement"
+      raise NotImplementedError, "Inheriting class must implement"
     end
 
     private