+++ spec/cloudscrape_client/executions/get_spec.rb
@@ -1,6 +1,8 @@
+# frozen_string_literal: true
+
 require "spec_helper"
 
-describe CloudscrapeClient::Executions::Get do
+RSpec.describe CloudscrapeClient::Executions::Get do
   let(:instance) { described_class.new(response: response) }
 
   let(:response) do