+++ spec/cloudscrape_client/executions/result_spec.rb
@@ -1,19 +1,31 @@
+# frozen_string_literal: true
+
 require "spec_helper"
 
-describe CloudscrapeClient::Executions::Result do
+RSpec.describe CloudscrapeClient::Executions::Result do
   let(:instance) do
     described_class.new(headers: response[:headers], row: response[:rows].first)
   end
 
   let(:response) do
     {
-      headers: %w(name age location avatars errors),
+      headers: %w(name age location avatars screenshot errors),
       rows: [rows]
     }
   end
 
-  let(:rows) { ["Chuck", 31, "Manchester", avatars, nil] }
+  let(:file) { "FILE:image/png;26071;11fed7f0-a508-4dc8-956a-481535c6f88a" }
   let(:avatars) { ["https://example.com/avatar1.png"] }
+  let(:rows) do
+    [
+      "Chuck",
+      31,
+      "Manchester",
+      avatars,
+      file,
+      nil
+    ]
+  end
 
   describe "#as_hash" do
     subject(:as_hash) { instance.as_hash }
@@ -24,6 +36,7 @@ describe CloudscrapeClient::Executions::Result do
         "age" => 31,
         "location" => "Manchester",
         "avatars" => avatars,
+        "screenshot" => file,
         "errors" => nil
       }
     end
@@ -62,6 +75,15 @@ describe CloudscrapeClient::Executions::Result do
 
     it "results avatars" do
       expect(avatars_method).to eq(avatars)
+    end
+  end
+
+  describe "#screenshot" do
+    subject(:screenshot_method) { instance.screenshot }
+
+    it "results ResultFile object" do
+      expect(screenshot_method)
+        .to be_an_instance_of(CloudscrapeClient::Executions::Result::File)
     end
   end